# docs/swagger.yaml
basePath: /
definitions:
  access.CheckInput:
    properties:
      action:
        example: read
        type: string
      resource:
        example: documents
        type: string
      tenant_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      user_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
    type: object
  access.CheckResponse:
    properties:
      allowed:
        type: boolean
    type: object
  action.Action:
    properties:
      app_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      metadata:
        type: object
      name:
        example: read
        type: string
      updated_at:
        type: string
    type: object
  action.CreateInput:
    properties:
      metadata:
        type: object
      name:
        example: read
        type: string
    type: object
  action.ListResponse:
    properties:
      actions:
        items:
          $ref: '#/definitions/action.Action'
        type: array
    type: object
  action.UpdateInput:
    properties:
      metadata:
        type: object
      name:
        example: read
        type: string
    type: object
  adminapikey.AdminAPIKey:
    properties:
      created_at:
        type: string
      id:
        type: string
      key:
        description: plaintext, only on create
        type: string
      name:
        type: string
      revoked_at:
        type: string
    type: object
  adminapikey.CreateInput:
    properties:
      name:
        type: string
    type: object
  adminapikey.ListResponse:
    properties:
      admin_api_keys:
        items:
          $ref: '#/definitions/adminapikey.AdminAPIKey'
        type: array
    type: object
...
  app.App:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      metadata:
        type: object
      name:
        example: acme
        type: string
      updated_at:
        type: string
    type: object
  app.CreateInput:
    properties:
      metadata:
        type: object
      name:
        example: acme
        type: string
    type: object
  app.ListResponse:
    properties:
      apps:
        items:
          $ref: '#/definitions/app.App'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    name: OpenAura
  description: |-
    Open-source user, role, and access management service.
    API version is selected via the X-API-Version header (not the URL path).
    App routes require an app API key (X-API-Key). Admin routes live under /admin and require an admin API key.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenAura API
  version: "1.0"
paths:
  /access/check:
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Access check request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/access.CheckInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/access.CheckResponse'
      summary: Check access
      tags:
      - access
  /actions:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/action.ListResponse'
      summary: List actions
      tags:
      - actions
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Action to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/action.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/action.Action'
      summary: Create action
      tags:
      - actions
  /actions/{id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Action ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete action
      tags:
      - actions
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Action ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/action.Action'
      summary: Get action
      tags:
      - actions
    patch:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Action ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/action.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/action.Action'
      summary: Update action
      tags:
      - actions
  /admin/api_keys:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adminapikey.ListResponse'
      summary: List admin API keys
      tags:
      - admin-api_keys
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Admin API key to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/adminapikey.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/adminapikey.AdminAPIKey'
      summary: Create admin API key
      tags:
      - admin-api_keys
  /admin/api_keys/{id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Admin API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Revoke admin API key
      tags:
      - admin-api_keys
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Admin API key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/adminapikey.AdminAPIKey'
      summary: Get admin API key
      tags:
      - admin-api_keys
  /admin/apps:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - default: 50
        description: Page size
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ListResponse'
      summary: List apps
      tags:
      - admin-apps
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: App to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/app.App'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      summary: Create app
      tags:
      - admin-apps
  /admin/apps/{id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: App ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete app
      tags:
      - admin-apps
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: App ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.App'
      summary: Get app
      tags:
      - admin-apps
    patch:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: App ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.App'
      summary: Update app
      tags:
      - admin-apps
  /admin/apps/{id}/api_keys:
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: Admin API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: App ID
        in: path
        name: id
        required: true
        type: string
      - description: API key to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apikey.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apikey.APIKey'
      summary: Create app API key (admin)
      tags:
      - admin-apps
  /api_keys:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apikey.ListResponse'
      summary: List app API keys
      tags:
      - api_keys
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: API key to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apikey.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/apikey.APIKey'
      summary: Create app API key
      tags:
      - api_keys
  /api_keys/{id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Revoke app API key
      tags:
      - api_keys
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apikey.APIKey'
      summary: Get app API key
      tags:
      - api_keys
  /resources:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resource.ListResponse'
      summary: List resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Resource to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resource.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resource.Resource'
      summary: Create resource
      tags:
      - resources
  /resources/{id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete resource
      tags:
      - resources
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resource.Resource'
      summary: Get resource
      tags:
      - resources
    patch:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resource.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resource.Resource'
      summary: Update resource
      tags:
      - resources
  /roleassignments:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by role ID
        in: query
        name: role_id
        type: string
      - description: Filter by tenant ID
        in: query
        name: tenant_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roleassignments.ListResponse'
      summary: List role assignments
      tags:
      - roleassignments
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Assignment to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/roleassignments.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/roleassignments.RoleAssignment'
      summary: Create role assignment
      tags:
      - roleassignments
  /roleassignments/{id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role assignment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete role assignment
      tags:
      - roleassignments
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role assignment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roleassignments.RoleAssignment'
      summary: Get role assignment
      tags:
      - roleassignments
    patch:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role assignment ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/roleassignments.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roleassignments.RoleAssignment'
      summary: Update role assignment
      tags:
      - roleassignments
  /roles:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/role.ListResponse'
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/role.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/role.Role'
      summary: Create role
      tags:
      - roles
  /roles/{id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete role
      tags:
      - roles
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/role.Role'
      summary: Get role
      tags:
      - roles
    patch:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/role.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/role.Role'
      summary: Update role
      tags:
      - roles
  /roles/{id}/permissions:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by resource ID
        in: query
        name: resource_id
        type: string
      - description: Filter by action ID
        in: query
        name: action_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permission.ListResponse'
      summary: List role permissions
      tags:
      - permissions
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Permission to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/permission.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/permission.Permission'
      summary: Create role permission
      tags:
      - permissions
  /roles/{id}/permissions/{permission_id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Permission ID
        in: path
        name: permission_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete role permission
      tags:
      - permissions
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Permission ID
        in: path
        name: permission_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permission.Permission'
      summary: Get role permission
      tags:
      - permissions
...
schemes:
- http
swagger: "2.0"
...